Improved Fortune Calculation
I have already put significant testing into this and believe it to be correct from the information available on the wiki regarding fortune.

I did learn that .getDrops() even on Lapis Ore, Redstone Ore, Glowstone, etc. will still only return 1. Even though Redstone Ore for example drops anywhere from 4-5 redstone (without Fortune). Because of this I had to build in a map of the min to max values. And for Glowstone and Redstone adjust the maximum based on Fortune. Also the bonus for Fortune on Glowstone can never exceed 4. So the maximum Glowstone Dust for Fortune 3 is 7.

Redstone and Glowstone dont follow the typical scaling math and only get the maximum drop amount increased based on fortune. The drop amount is then randomly picked from the minimum and the new maximum value.

For all other ores a starting amount is picked based on the minimum and maximum values (mainly only applies to lapis) as the min and max for everything besides lapis is 1.

Then it rolls a number from 0-100 and iterates through the multipliers to figure out the new amount. An example of how this math works on Fortune 3 is explained below. The chance for the multiple to be 1 (drop amount unchanged) is weighted as 2 and all other multiples has a weight of 1.

Fortune 3 has 4 drop tiers (enchLevel + 1).
These tiers are 1, 2, 3, and 4. This becomes the multiplier and the drop amount is multipled by the multiplier. Tier 1 is dropping 1 (or unchanging the drop amount for Lapis). Tier 2 multiplies the drop amount by 2, Tier 3 by 3, etc., etc.

Fortune 3 has a 40% chance to land in Tier 1. The remaining 60% is split up between Tier 2, 3, and 4 for a 20% chance for each tier after 1.

In this example if the roll <= 40 the multiplier will be 1.
lf the roll is 41-60 the multplier will be 2, 61-80 will be multiplier 3, and 81-100 the multiplier will be 4.

A debug boolean is including to toggle the output of the math on these calculations out to the console.

This should be a very accurate representation of Fortune in vanilla (assuming there are no bugs).

I also removed the need for the NMS code that was there previously. This should work on all and future minecraft versions and wont crash due to a bad cast.