Command Improvements, various bug fixes and improvements
Added proper checking on command parameters to help users who may accidentally type the command wrong. Instead of saying nothing we now tell them the error and show the proper command usage.

Fixed a major bug in the full inventory detection. Instead of checking the inventorys max stack size we check the item slots max stack size. This prevents the plugin thinking things such as Diamond Pickaxes (which only can stack to 1) stack to 64. This now works properly for those items and will no longer delete the pickaxe if the player inventory was full and will now drop it on the ground as originally intended. Also added a warning to both the sender and receiver that the item was dropped to the ground as the inventory was full.

Based the "Pickaxe Repaired" chat message color off the pickaxe type just as the original plugin did.

Added the Full Inventory sound to the blocks sound category so the volume can be adjusted.